when copying a scenario, automatically pop up the edit name modal; fixed small bug with editing input
